(chore) Add FDL docs to the service manual
The documentation for FDL is currently sitting in a Google Doc, and we'd
like to move it into the service manual, the canonical place for docs to
live. This is a copy of the current contents of the Google Doc with no
modifications and retaining the bulk of the formatting.